gpio/requirements.txt
pyyaml
dht11